edit delete from fav

+++ MoviesList_Swift/Services/FavoriteCoreDao.swift
@@ -89,51 +89,68 @@ class FavoriteCoreDao {
         let appDelegate = UIApplication.shared.delegate as! AppDelegate
         let context = appDelegate.persistentContainer.viewContext
         
-        let movie = FavoriteMovie(context: context)
-        //NSEntityDescription.insertNewObject(forEntityName: "FavoriteMovie", into: context) as! FavoriteMovie
-        movie.id_movie = newMovie.id!
-        movie.original_title = newMovie.original_title
-        movie.poster_path = newMovie.poster_path
-        movie.overview = newMovie.overview
-        movie.vote_average = newMovie.vote_average!
-        movie.release_date = newMovie.release_date
-        
-        for newreview in newMovie.reviewsArray!
-        {
-            let review = FavoriteReview(context: context)
-            //NSEntityDescription.insertNewObject(forEntityName: "FavoriteReview", into: context) as! FavoriteReview
-           
-            review.id_Review = newreview.id
-            review.author = newreview.author
-            review.content = newreview.content
-            review.url = newreview.url
-            
-            movie.addToHasReview(review)
-            review.aboutMovie = movie
-        }
-
-        for newVideo in newMovie.videosArray!
-        {
-            let video = FavoriteVideo(context: context)
-            //NSEntityDescription.insertNewObject(forEntityName: "FavoriteVideo", into: context) as! FavoriteVideo
-            video.name  = newVideo.name
-            video.id_video = newVideo.id
-            video.iso_3166_1 = newVideo.iso_3166_1
-            video.iso_639_1 = newVideo.iso_639_1
-            video.key = newVideo.key
-            video.site = newVideo.site
-         //   video.size = newVideo.size as! NSDecimalNumber
-            
-            movie.addToHasVideo(video)
-            video.aboutMovie = movie
-        }
-        
+        let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: "FavoriteMovie")
+        let myPredicate = NSPredicate(format: "id_movie==\( newMovie.id!)")
+        fetchRequest.predicate = myPredicate
         do{
-            try context.save()
+            if(try context.fetch(fetchRequest).capacity == 0)
+            {
+                let movie = FavoriteMovie(context: context)
+                //NSEntityDescription.insertNewObject(forEntityName: "FavoriteMovie", into: context) as! FavoriteMovie
+                movie.id_movie = newMovie.id!
+                movie.original_title = newMovie.original_title
+                movie.poster_path = newMovie.poster_path
+                movie.overview = newMovie.overview
+                movie.vote_average = newMovie.vote_average!
+                movie.release_date = newMovie.release_date
+                
+                for newreview in newMovie.reviewsArray!
+                {
+                    let review = FavoriteReview(context: context)
+                    //NSEntityDescription.insertNewObject(forEntityName: "FavoriteReview", into: context) as! FavoriteReview
+                    
+                    review.id_Review = newreview.id
+                    review.author = newreview.author
+                    review.content = newreview.content
+                    review.url = newreview.url
+                    
+                    movie.addToHasReview(review)
+                    review.aboutMovie = movie
+                }
+                
+                for newVideo in newMovie.videosArray!
+                {
+                    let video = FavoriteVideo(context: context)
+                    //NSEntityDescription.insertNewObject(forEntityName: "FavoriteVideo", into: context) as! FavoriteVideo
+                    video.name  = newVideo.name
+                    video.id_video = newVideo.id
+                    video.iso_3166_1 = newVideo.iso_3166_1
+                    video.iso_639_1 = newVideo.iso_639_1
+                    video.key = newVideo.key
+                    video.site = newVideo.site
+                    //   video.size = newVideo.size as! NSDecimalNumber
+                    
+                    movie.addToHasVideo(video)
+                    video.aboutMovie = movie
+                }
+                
+                do{
+                    try context.save()
+                }
+                catch {
+                    print("Error")
+                }
+                
+                
+            }
         }
-        catch {
-            print("Error")
+            
+        catch
+        {
+            print("Not deleted!")
         }
+        
+        
     }
     
     func deleteFromFavoriteMovies(newMovie: Movie){
@@ -144,10 +161,15 @@ class FavoriteCoreDao {
         let myPredicate = NSPredicate(format: "id_movie==\( newMovie.id!)")
         fetchRequest.predicate = myPredicate
         
+       
         do{
-            let deletedObj = try context.fetch(fetchRequest)[0] as! NSManagedObject
-            context.delete(deletedObj)
-        }
+            if(try context.fetch(fetchRequest).capacity != 0)
+            {
+                let deletedObj = try context.fetch(fetchRequest)[0] as! NSManagedObject
+                context.delete(deletedObj)
+            }
+            }
+           
         catch
         {
             print("Not deleted!")
